Tests: Started adding full tests for WP Super Cache, WP-Optimize, WP Fastest Cache

+++ tests/_support/Helper/Acceptance/WPCachePlugins.php
@@ -69,21 +69,14 @@ class WPCachePlugins extends \Codeception\Module
 	}
 
 	/**
-	 * Helper method to activate the WP Fastest Cache Plugin, and then
-	 * enable it to perform caching.
+	 * Helper method to enable caching in the WP Fastest Cache Plugin.
 	 *
 	 * @since   2.2.2
 	 *
 	 * @param   AcceptanceTester $I      Acceptance Tester.
 	 */
-	public function activeAndEnableWPFastestCachePlugin($I)
+	public function enableCachingWPFastestCachePlugin($I)
 	{
-		// Clear up any cache configuration files that might exist from previous tests.
-		$I->deleteAdvancedCacheConfig($I);
-
-		// Activate Plugin.
-		$I->activateThirdPartyPlugin($I, 'wp-fastest-cache');
-
 		// Navigate to its settings screen.
 		$I->amOnAdminPage('admin.php?page=wpfastestcacheoptions');
 
@@ -92,24 +85,23 @@ class WPCachePlugins extends \Codeception\Module
 
 		// Save.
 		$I->click('Submit');
+
+		// The ConvertKit Plugin will now automatically add an exclusion rule
+		// to WP Fastest Cache. Check this rule does exist in the settings.
+		$I->amOnAdminPage('admin.php?page=wpfastestcacheoptions');
+		$I->click('label[for="wpfc-exclude"]');
+		$I->see('Contains: ck_subscriber_id');
 	}
 
 	/**
-	 * Helper method to activate the WP Optimize Plugin, and then
-	 * enable it to perform caching.
+	 * Helper method to enable caching in the WP Optimize Plugin.
 	 *
 	 * @since   2.2.2
 	 *
 	 * @param   AcceptanceTester $I      Acceptance Tester.
 	 */
-	public function activeAndEnableWPOptimizePlugin($I)
+	public function enableCachingWPOptimizePlugin($I)
 	{
-		// Clear up any cache configuration files that might exist from previous tests.
-		$I->deleteAdvancedCacheConfig($I);
-
-		// Activate Plugin.
-		$I->activateThirdPartyPlugin($I, 'wp-optimize');
-
 		// Navigate to its settings screen.
 		$I->amOnAdminPage('admin.php?page=wpo_cache');
 
@@ -125,21 +117,14 @@ class WPCachePlugins extends \Codeception\Module
 	}
 
 	/**
-	 * Helper method to activate the WP Super Cache Plugin, and then
-	 * enable it to perform caching.
+	 * Helper method to enable caching in the WP Super Cache Plugin.
 	 *
 	 * @since   2.2.2
 	 *
 	 * @param   AcceptanceTester $I      Acceptance Tester.
 	 */
-	public function activeAndEnableWPSuperCachePlugin($I)
+	public function enableCachingWPSuperCachePlugin($I)
 	{
-		// Clear up any cache configuration files that might exist from previous tests.
-		$I->deleteAdvancedCacheConfig($I);
-
-		// Activate Plugin.
-		$I->activateThirdPartyPlugin($I, 'wp-super-cache');
-
 		// Navigate to its settings screen.
 		$I->amOnAdminPage('options-general.php?page=wpsupercache');
 
@@ -151,16 +136,41 @@ class WPCachePlugins extends \Codeception\Module
 	}
 
 	/**
-	 * Helper method to delete the file at wp-content/advanced-cache.php, which
-	 * may have been created by a previous caching plugin that was enabled
-	 * in a previous test.
+	 * Helper method to exclude caching when a cookie is present
+	 * in the WP Super Cache Plugin.
+	 *
+	 * @since   2.2.2
+	 *
+	 * @param   AcceptanceTester $I      Acceptance Tester.
+	 */
+	public function excludeCachingWPSuperCachePlugin($I, $cookieName = 'ck_subscriber_id')
+	{
+		// Navigate to its settings screen.
+		$I->amOnAdminPage('options-general.php?page=wpsupercache&tab=settings');
+
+		// Add cookis to "Rejected Cookies" setting.
+		$I->fillField('wp_rejected_cookies', $cookieName);
+
+		// Save.
+		$I->click('form[name="wp_edit_rejected_cookies"] input.button-primary');
+	}
+
+	/**
+	 * Helper method to delete the files at wp-content/advanced-cache.php
+	 * and wp-content/wp-cache-config.php, which may have been created by a
+	 * previous caching plugin that was enabled in a previous test.
 	 * 
 	 * @since 	2.2.2
 	 *
 	 * @param   AcceptanceTester $I      Acceptance Tester.
 	 */
-	public function deleteAdvancedCacheConfig($I)
+	public function deleteWPCacheConfigFiles($I)
 	{
-		$I->deleteFile($_ENV['WP_ROOT_FOLDER'] . '/wp-content/advanced-cache.php');
+		if (file_exists($_ENV['WP_ROOT_FOLDER'] . '/wp-content/advanced-cache.php')) {
+			$I->deleteFile($_ENV['WP_ROOT_FOLDER'] . '/wp-content/advanced-cache.php');
+		}
+		if (file_exists($_ENV['WP_ROOT_FOLDER'] . '/wp-content/wp-cache-config.php')) {
+			$I->deleteFile($_ENV['WP_ROOT_FOLDER'] . '/wp-content/wp-cache-config.php');
+		}
 	}
 }
